feat: login prompt component
