+++ src/index.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import Home from 'pages/Home';
 import DetailMenu from 'pages/DetailMenu';
 import Notification from 'pages/Notification';
 import OrderComplete from 'pages/OrderComplete';
@@ -10,6 +9,7 @@ import MenuList from 'pages/MenuList';
 import configureStore from 'reduxes';
 import reactDom from 'react-dom';
 import * as firebase from 'firebase';
+import PromptLogin from 'pages/PromptLogin';
 import * as serviceWorker from './serviceWorker';
 
 export const providerGoogle = new firebase.auth.GoogleAuthProvider();
@@ -32,13 +32,14 @@ const store = configureStore();
 const App = () => (
 
   <Switch>
-    <Route exact path="/" component={Home} />
     <Route exact path="/detail" component={DetailMenu} />
     <Route exact path="/notification" component={Notification} />
     <Route exact path="/ordered" component={OrderComplete} />
     <Route exact path="/paid" component={PaymentComplete} />
     <Route exact path="/menus" component={MenuList} />
     <Route exact path="/notif" component={Notification} />
+    {/* 必要ないけど確認するためにルーティング追加してる */}
+    <Route exact path="/prompt" component={PromptLogin} />
   </Switch>
 
 );
